Separated mkdir*(), so they can be required without fs-extra
